feat: move the `main.py` in the `cli` directory
